+++ Nanoforge/util/MathUtil.h
@@ -5,10 +5,10 @@
 
 static f32 ToRadians(f32 angleDegrees)
 {
-    return angleDegrees * (M_PI / 180.0f);
+    return angleDegrees * ((f32)M_PI / 180.0f);
 }
 
 static f32 ToDegrees(f32 angleRadians)
 {
-    return angleRadians * (180.0f / M_PI);
+    return angleRadians * (180.0f / (f32)M_PI);
 }
